% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_y.R
\name{causal_y}
\alias{causal_y}
\title{Estimate Causal Effect of an Intervention}
\usage{
causal_y(Sigma_D, A, i, X_mat, g)
}
\arguments{
\item{Sigma_D}{A \eqn{q,q} covariance matrix, typically a sample from the posterior distribution.}

\item{A}{A \eqn{q,q} adjacency matrix.}

\item{i}{An integer representing the label of the node being intervened upon (i in {2, ..., q}).}

\item{X_mat}{An \eqn{n,q-1} data matrix containing the observations for the non-response variables (nodes 2 through q).}

\item{g}{A numeric hyperparameter.}
}
\value{
A numeric vector of \eqn{n,1}, where each element represents the estimated causal effect (probability) for the corresponding observation in \code{X_mat}.
}
\description{
This function calculates the individual causal effects of performing an intervention
\code{do(X_i = x_i)} on a binary response variable (node 1). The effect is estimated as the probability
\eqn{Pr(Y = 1 | do(X_i = x_i), Sigma_D, x_i)}.
}
