% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianRankCorr.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert a correlation matrix to a covariance matrix}
\usage{
cor2cov(corr, std)
}
\arguments{
\item{corr}{The correlation matrix to be converted. This must be symmetric.}

\item{std}{A vector that contains the standard deviations of the variables in the correlation matrix.}
}
\value{
The covariance matrix.
}
\description{
This function converts a correlation matrix to a covariance matrix
}
