% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{dd}
\alias{dd}
\title{Delete a Directed Edge}
\usage{
dd(A, nodes)
}
\arguments{
\item{A}{A \eqn{q,q} adjacency matrix.}

\item{nodes}{A numeric vector of \eqn{(2,1)}, \code{c(x, y)}, representing the nodes.}
}
\value{
The modified adjacency matrix.
}
\description{
Deletes the directed edge from node x to node y in the adjacency matrix.
}
\keyword{internal}
