% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSL_model.R
\name{fn}
\alias{fn}
\title{Functions to be used in bsl (for internal use)}
\usage{
fn(.Object)
}
\arguments{
\item{.Object}{A ``BSLMODEL'' object.}
}
\description{
Generate the generic simulation and summary statistics function for \code{n} simulations
and fixed \code{theta} (for internal use).
}
