% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BSL.R
\name{getLoglike}
\alias{getLoglike}
\title{Obtain the log-likelihoods from a "BSL" object}
\usage{
getLoglike(object, ...)
}
\arguments{
\item{object}{A ``BSL'' class object.}

\item{...}{Other arguments.}
}
\value{
The vector of log likelihood evaluations, after removing burn-in and thinning.
}
\description{
see \code{\link{BSLclass}}
}
