% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-PENALTY.R
\name{getPenalty}
\alias{getPenalty}
\title{Obtain the selected penalty values from a "PENALTY" object}
\usage{
getPenalty(object, ...)
}
\arguments{
\item{object}{A ``PENALTY'' class object.}

\item{...}{Other arguments.}
}
\value{
The selecty penalty values.
}
\description{
see \code{\link{PENALTYclass}}
}
