% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmmScores}
\alias{gmmScores}
\title{This function computes the scores of Gaussian mixture models for the toad example}
\usage{
gmmScores(X, gmm, lag)
}
\arguments{
\item{X}{The observation matrix.}

\item{gmm}{A matrix containing the parameters for a Gaussian mixture model.
The first row should be component proportions. The second row should be
component means. The last row should be component Variances.}

\item{lag}{The lag of days to compute the displacements.}
}
\value{
A list of the following vectors: return frequencies, scores of
component proportion, scores of mean and scores of variance.
}
\description{
The summary statistics for the toad example is taken to be the
scores of Gaussian mixture model fitted to the displacement distribution. The
displacements are computed with a given day lag.
}
