% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_1.R
\name{m_1}
\alias{m_1}
\title{Computes marginal likelihood for node 1}
\usage{
m_1(dag, X, n, a, g)
}
\arguments{
\item{dag}{A \eqn{q,q} adjacency matrix representing the DAG structure.}

\item{X}{An \eqn{n,q} data matrix of observations.}

\item{n}{number of observations (rows) in the data matrix \code{X}.}

\item{a}{A hyperparameter of the prior distribution.}

\item{g}{A hyperparameter of the prior distribution.}
}
\value{
The log of the marginal likelihood for node 1.
}
\description{
Computes the log marginal likelihood for node 1 (the 0-1 response variable) in a DAG,
given the data X. This function uses a different formula tailored for the Probit
model's response variable.
}
