% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_j.R
\name{m_j}
\alias{m_j}
\title{Computes node-marginal likelihood of a DAG model (j in {2,...,q})}
\usage{
m_j(j, dag, X, n, a, g)
}
\arguments{
\item{j}{An integer representing the label of the \code{j} in \code{dag} (must be > 1).}

\item{dag}{A \eqn{q,q} adjacency matrix representing the \code{dag} structure.}

\item{X}{An \eqn{n,q} data matrix of observations.}

\item{n}{number of observations (rows) in the data matrix \eqn{X}.}

\item{a}{A hyperparameter of the prior distribution.}

\item{g}{A hyperparameter of the prior distribution, related to the prior variance.}
}
\value{
The logarithm of the marginal likelihood for the specified \code{j}.
}
\description{
This function computes the logarithm marginal likelihood for a specific node j (where j is in {2,...,q}) in a DAG,
given the data X. This calculation is based on a DAG-Probit model with a Normal-Gamma
prior distribution and utilizes a Laplace approximation for calculating marginal likelihood.
}
\details{
The function handles two cases: when the node has no parents and when it has one or more parents.
The calculation involves terms from the Normal-Gamma prior and approximations based on the data.
The formula for \code{a} hyperparameter has been adjusted to be \code{nrow(X)-1}.
}
