% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_like_j.R
\name{marg_like_j}
\alias{marg_like_j}
\title{Compute log marginal likelihood for a Node in a DAG}
\usage{
marg_like_j(j, dag, X, a, g, n)
}
\arguments{
\item{j}{An integer representing the label of the \code{j} for which the likelihood is calculated.}

\item{dag}{A \eqn{q,q} adjacency matrix representing the DAG structure.}

\item{X}{An \eqn{n,q} data matrix of observations.}

\item{a}{A hyperparameter of the prior distribution.}

\item{g}{A hyperparameter of the prior distribution.}

\item{n}{number of observations (rows) in the data matrix \eqn{X}.}
}
\value{
A single numeric value: the log of the marginal likelihood for the specified node \code{j}.
}
\description{
A wrapper function that calculates the log marginal likelihood for a specified node
in a DAG. It automatically determines whether to use the formula for the general
node or the special formula for the response node (node 1).
}
