% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_dag_probit.R
\name{mcmc_dag_probit}
\alias{mcmc_dag_probit}
\title{MCMC Sampler for DAG-Probit Models}
\usage{
mcmc_dag_probit(y, X, TT, burn, a = NULL, g = NULL, w = NULL, causal = TRUE)
}
\arguments{
\item{y}{An \eqn{n,1} numeric vector of observations from the binary response variable.}

\item{X}{An \eqn{n,p} numeric matrix of observations for the p covariates.}

\item{TT}{An integer specifying the total number of MCMC iterations.}

\item{burn}{An integer specifying the number of burn-in iterations to be discarded.}

\item{a}{A numeric hyperparameter for the prior distribution. If NULL, it defaults to q.}

\item{g}{A numeric hyperparameter for the prior distribution. If NULL, it defaults to 1/n.}

\item{w}{A numeric prior probability of edge inclusion. If NULL, it defaults to 0.5.}

\item{causal}{A logical value. If TRUE, BMA causal effects are computed; if FALSE, they are not.}
}
\value{
A list containing the MCMC results:
\itemize{
\item \code{A_chain}: matrix with TT columns, each representing the (vectorized) \eqn{q,q} adjacency matrix of one visited \code{DAG}
\item \code{Sigma_chain}: matrix with TT columns, each representing the (vectorized) \eqn{q,q} covariance matrix of one visited \code{DAG}
\item \code{Causal_hat}: \eqn{n,q} matrix with BMA estimates of subject-specific causal effects (one row for each subject, one column for each intervened node)
}
}
\description{
This function implements a Markov Chain Monte Carlo (MCMC) algorithm to learn the structure
of a Directed Acyclic Graph (DAG) from data, assuming a probit model for the binary
response variable.
It can also compute Bayesian Model Averaged (BMA) estimates of subject-specific causal effects.
}
