% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Perform local moves given a DAG}
\usage{
move(A, q = q)
}
\arguments{
\item{A}{A \eqn{q,q} adjacency matrix of the current \code{DAG}.}

\item{q}{The number of \code{nodes} in the \code{DAG}.}
}
\value{
A list containing three elements:
\itemize{
\item \code{A_new}: The adjacency matrix after the valid move has been applied.
\item \code{type.operator}: An integer \eqn{{1,2,3}} indicating the type of move performed (1: insert, 2: delete, 3: reverse).
\item \code{nodes}: A numeric vector of \eqn{(2,1)}, indicating the pair of nodes involved in the move.
}
}
\description{
This function locally proposes a DAG by inserting (\code{id}), deleting (\code{dd}) or reversing (\code{rd}) an edge between two \code{nodes} on a given DAG's adjacency matrix.
It randomly selects a possible move from a predefined set of actions, applies it, and checks if the resulting graph is still a DAG.
The process is repeated until a valid move that results in a DAG is found.
}
\details{
The function defines three types of moves internally:
\describe{
\item{1 (Insertion)}{Adds a directed edge from node x to node y.}
\item{2 (Deletion)}{Deletes the existing directed edge from node x to node y.}
\item{3 (Reversal)}{Deletes the edge from node x to node y and adds a new edge from node y to node x.}
}
This function relies on helper functions \code{id()}, \code{dd()}, and \code{rd()}, as well as a function called \code{is.DAG()} to validate the graph structure.
}
