% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMiniProgressBar.R
\name{myMiniProgressBar}
\alias{myMiniProgressBar}
\title{Progress Bar}
\usage{
myMiniProgressBar(p, txt1 = "", txt2 = "", style = 1, label = c("=", "-", "|"))
}
\arguments{
\item{p}{Numeric, percentage of finished progress, between 0 and 1.}

\item{txt1}{String to put before the progress bar}

\item{txt2}{String to put after the progress bar}

\item{style}{The display style. 1 is single-lined; 2 is double-lined; 3 display the progress in a 5-lined block.}

\item{label}{Character labels for "finished", "un-finished", and "side bars".}
}
\value{
No return value, called for side effects.
}
\description{
Print a customisable progress bar in the console.
}
\keyword{internal}
