% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.edge.R
\name{n.edge}
\alias{n.edge}
\title{Count the Number of Edges in a DAG}
\usage{
n.edge(A)
}
\arguments{
\item{A}{A \eqn{q,q} adjacency matrix of the DAG.}
}
\value{
An integer representing the total number of edges.
}
\description{
This function counts the number of directed edges in a DAG represented by an adjacency matrix.
}
\keyword{internal}
