% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obsMat2deltaxR}
\alias{obsMat2deltaxR}
\title{Convert an observation matrix to a vector of n-day displacements}
\usage{
obsMat2deltaxR(X, lag)
}
\arguments{
\item{X}{The observation matrix to be converted.}

\item{lag}{Interger, the number of day lags to compute the displacement.}
}
\value{
A vector of displacements.
}
\description{
Convert an observation matrix to a vector of n-day displacements
}
