% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_penbsl.R
\name{penbsl}
\alias{penbsl}
\alias{print.penbsl}
\alias{plot.penbsl}
\title{S3 reference class of the result from tuning to select the optimal penalty for BSLasso}
\usage{
\method{print}{penbsl}(x, digits = max(3L, getOption("digits") - 4L))

plot.penbsl(x, logscale = TRUE)
}
\arguments{
\item{x}{A "penbsl" class object, typically the output of function \code{\link{selectPenalty}}.}

\item{digits}{The number of digits to print.}

\item{logscale}{A logical indicator whether the x-axis (penalty) should be log transformed. The
default is \code{TRUE}.}
}
\description{
Two functions (print and plot) are provided for class "penbsl".
}
\author{
Ziwen An, Christopher C. Drovandi and Leah F. South
}
