% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sigma.R
\name{posterior_sigma}
\alias{posterior_sigma}
\title{Sample from the Posterior Distribution of the Covariance Matrix}
\usage{
posterior_sigma(Y, A, g, a)
}
\arguments{
\item{Y}{An \eqn{n,q} data matrix of observations.}

\item{A}{A \eqn{q,q} adjacency matrix representing the \code{DAG} structure.}

\item{g}{A hyperparameter of the prior distribution.}

\item{a}{A hyperparameter of the prior distribution.}
}
\value{
A list containing three elements:
\itemize{
\item \code{Sigma_post}: A \eqn{q,q} matrix, which is a single sample from the posterior distribution of the covariance matrix Sigma.
\item \code{L_post}: A \eqn{q,q} lower-triangular matrix, the Cholesky factor \code{L} from the reparameterization.
\item \code{D_post}: A \eqn{q,q} diagonal matrix, the diagonal matrix \code{D} from the reparameterization.
}
}
\description{
This function generates a single sample from the posterior distribution of the
covariance matrix Sigma, given a data matrix \code{Y} and a \code{DAG} structure A. The sampling
is performed by first sampling from the posterior of the Cholesky decomposition
parameters (\code{L} and \code{D}).
}
