% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_y_to_binary.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for DAG-Probit Model}
\usage{
prepare_data(Y, cutoff = 0)
}
\arguments{
\item{Y}{An \eqn{n,q} data matrix where the first column is a continuous variable
that will be converted to a binary response.}

\item{cutoff}{A numeric value. Values greater than the cutoff are set to 1,
and values less than or equal to the cutoff are set to 0. Defaults to 0.}
}
\value{
A list containing two elements:
\itemize{
\item \code{y}: An \eqn{n,1} numeric vector of the binary response variable.
\item \code{X}: An \eqn{n,q-1} numeric matrix of covariates.
}
}
\description{
This function takes a data matrix \code{Y} and preprocesses it for the DAG-probit model.
It separates the first column as the binary response variable \code{y} and the remaining
columns as the covariate matrix \code{X}. The response variable is then converted
\code{to binary (0/1)} based on a cutoff value.
}
