% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{rd}
\alias{rd}
\title{Reverse a Directed Edge}
\usage{
rd(A, nodes)
}
\arguments{
\item{A}{A \eqn{q,q} adjacency matrix.}

\item{nodes}{A numeric vector of \eqn{(2,1)}, \code{c(x, y)}, representing the nodes.}
}
\value{
The modified adjacency matrix.
}
\description{
Reverses the direction of an edge from x -> y to y -> x in the adjacency matrix.
}
\keyword{internal}
