% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rstable}
\alias{rstable}
\title{Generate a random sample from the zero-centered stable distribution}
\usage{
rstable(scale, alpha)
}
\arguments{
\item{scale}{The scale parameter.}

\item{alpha}{The stability parameter.}
}
\value{
A random sample from the zero-centered stable distribution.
}
\description{
Draw a sample from a symmetric, zero-centered stable distribution with
given scale and stability (alpha) parameters, using the CMS algorithm.
}
\keyword{internal}
