% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSL_model.R
\name{setns}
\alias{setns}
\title{Find the length of summary statistics (for internal use)}
\usage{
setns(.Object)
}
\arguments{
\item{.Object}{A ``BSLMODEL'' object.}
}
\description{
Find and set the length of summary statistics with a test run (for internal use).
}
