% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{show,bsl-method}
\alias{show,bsl-method}
\title{Show method for class "bsl". Display the basic information of a bsl object.}
\usage{
\S4method{show}{bsl}(object)
}
\description{
Show method for class "bsl". Display the basic information of a bsl object.
}
