% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Functions.R
\name{show.bsl}
\alias{show.bsl}
\title{Show method for class "bsl". Display the basic information of a bsl object.}
\usage{
show.bsl(object)
}
\arguments{
\item{object}{A "bsl" class object to be displayed.}
}
\description{
Display the basic information of a bsl object.
}
