% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-MODEL.R
\name{simulation}
\alias{simulation}
\title{Run simulations with a give "MODEL" object}
\usage{
simulation(model, ...)
}
\arguments{
\item{model}{A ``MODEL'' object.}

\item{...}{Other arguments.}
}
\description{
see \code{\link{MODEL}}
}
