% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-MODEL.R
\name{summStat}
\alias{summStat}
\title{Compute the summary statistics with the given data}
\usage{
summStat(x, model)
}
\arguments{
\item{x}{The data to pass to the summary statistics function.}

\item{model}{A ``MODEL'' object.}
}
\description{
see \code{\link{MODEL}}
}
