% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BSL.R
\name{getGamma}
\alias{getGamma}
\title{Obtain the gamma samples (the latent parameters for BSLmisspec method) from a
"BSL" object}
\usage{
getGamma(object, ...)
}
\arguments{
\item{object}{A ``BSL'' class object.}

\item{...}{Other arguments.}
}
\description{
see \code{\link{BSLclass}}
}
