% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BSL.R
\name{getTheta}
\alias{getTheta}
\title{Obtain the samples from a "BSL" object}
\usage{
getTheta(object, ...)
}
\arguments{
\item{object}{A ``BSL'' class object.}

\item{...}{Other arguments.}
}
\description{
see \code{\link{BSLclass}}
}
